requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements.txt requirements.in
#
django==1.11.4
djangorestframework==3.6.3
gunicorn==19.7.1
psycopg2==2.7.3
pytz==2017.2              # via django
